# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RENOVA redevelopment operating platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: string;
  name: string;
  location: string;
  societyType: string;
  landType?: string;
  stage: string;
  readiness: number;
  memberCount?: number;
  updatedAt: string;
  redFlags: string[];
  nextStep: string;
  pathway?: string[];
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  location: string;
  /** @minLength 1 */
  societyType: string;
  landType?: string;
  /** @minimum 1 */
  memberCount?: number;
}

export interface RequirementInput {
  /** @minLength 2 */
  societyName: string;
  /** @minLength 2 */
  location: string;
  /** @minLength 1 */
  societyType: string;
  /** @minimum 1 */
  memberCount: number;
  /** @minimum 1 */
  buildingAge: number;
  landType?: string;
  /** @minimum 1 */
  plotArea?: number;
  conveyanceStatus?: string;
  structuralAudit?: string;
  /** @minItems 1 */
  services: string[];
  /** @minLength 1 */
  timeline: string;
  /** @maxLength 1000 */
  brief?: string;
  /** @minLength 2 */
  contactName: string;
  /** @minLength 1 */
  contactRole: string;
  /** @minLength 10 */
  phone: string;
  /** @minLength 3 */
  email: string;
  consent: boolean;
}

export interface RequirementSubmission {
  id: string;
  reference: string;
  status: string;
  submittedAt: string;
  projectId: string;
  nextStep: string;
}

export interface Activity {
  id: string;
  title: string;
  detail: string;
  date: string;
  kind: string;
}

export interface Dashboard {
  activeProject: Project;
  totalReadiness: number;
  openRedFlags: number;
  completedSteps: number;
  recentActivity: Activity[];
}

export interface Professional {
  id: string;
  name: string;
  role: string;
  location: string;
  experience: string;
  match: number;
  specialties: string[];
  verified: boolean;
}

export interface Regulation {
  id: string;
  code: string;
  title: string;
  summary: string;
  authority: string;
  updatedAt: string;
  status: string;
}

export type ListProfessionalsParams = {
role?: string;
};