# lib/api-zod/src/generated/types/requirementSubmission.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RENOVA redevelopment operating platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RequirementSubmission {
  id: string;
  reference: string;
  status: string;
  submittedAt: string;
  projectId: string;
  nextStep: string;
}